[refactor] refactor import models, services
