+++ backend/src/models/email/index.ts
@@ -1,0 +1,1 @@
+export * from './confirmationEmail';
