+++ backend/src/models/food/index.ts
@@ -1,0 +1,4 @@
+export * from './dietRecord';
+export * from './ingredient';
+export * from './recipe';
+export * from './recipeIngredientRelation';
