+++ backend/src/models/index.ts
@@ -1,0 +1,5 @@
+export * from './email';
+export * from './food';
+export * from './guild';
+export * from './notification';
+export * from './user';
