+++ backend/src/services/guild/guild.ts
@@ -1,7 +1,6 @@
 import { ApplicationError } from '../../utils/error/applicationError';
 import { BaseGuild, Guild } from '../../types/guild/guild';
-import { UserGuildRelationModel } from '../../models/user/userGuildRelation';
-import { GuildModel } from '../../models/guild/guild';
+import { UserGuildRelationModel, GuildModel } from '../../models';
 
 const defaultTitle = 'Personal Cabin';
 const defaultDescription = `In your Personal Cabin, you have the flexibility to select from various mission types like 'Ordinary', 'Emergency', and 'Repetitive', tailoring them to your specific needs. Additionally, you can customize the recurrence frequency, whether it's daily, weekly, or monthly, to suit your schedule. Furthermore, missions can be further broken down into multiple sub-goals, empowering you to gain a comprehensive overview of your pending missions and strategize your approach for more efficient planning and completion.`;
